cool#8023 browser, clipboard: try using navigator.clipboard before showing the paste popup
When pasting with Ctrl-V, paste worked already. When pasting with the
paste button, we just opened a popup, telling the user to do Ctrl-V
instead.

The reason for this is that we used to paste using execCommand('paste'),
which typically fails in modern browsers (reportedly not in IE11).

Use the new navigator.clipboard.read() API (available in Chrome, not in
Firefox) to also actually do a proper HTML paste when using the toolbar
button.

Note that the new API strips away <meta> tags, so this depends on
previous work to switch to a <div id="meta-origin"
data-coolorigin="...">...</div> marker around the entire HTML. This also
just implements HTML paste, plain text and images still need adding;
paste special is unchanged for now.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I2a9340719e99695131f9caa452b5fbfb53573e6c