cool#8023 clipboard: remove old marker
The old marker is stripped away when Chrome sanitizes the incoming
clipboard HTML. We used to go with both markers to allow an incremental
conversion.

By now, there are no remaining consumers of the old marker, so remove it
before somebody would depend on both markers by accident.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Ib9dd0e7cf64e1a99d6038a179651ef5e114c6f8b