calc: on tab swithing do not replay print twips messages
From CanvasTileLayer._saveMessageForReplay description:
"We will not get some messages (with coordinates) from core when zoom
changes because print-twips coordinates are zoom-invariant. So we need
to remember the last version of them and replay, when zoom is changed.
In calc we need to replay the messages when sheet-geometry changes
too. This is because it is possible for the updated print-twips
messages to arrive before the sheet-geometry update message arrives."

The problem is that after switching to a new tab it could occur that
some of these stored messages are replayed on the new sheet even if
they were related to the previous one.
So for instance if in the old tab there is a cell selection we could
end up try to select such cells in the new sheet too, and even if such
selection is successively cleared by further message specific for the
new sheet, such message replaya was causing an unexpected document
scroll.

Signed-off-by: Marco Cecchetti <marco.cecchetti@collabora.com>
Change-Id: I63995aaa51c25216abd246551fbd2ec437a3f6de