Allow unfollowing by clicking the followed user
Previously this would do nothing (as it tried to follow the user
again). This was bad, as it was a struggle to figure out how to
unfollow someone. Now clicking on the followed user again unfollows
them, which mirrors the functionality in other apps such as Figma.

Happily, as the previous action appeared as a no-op to the user, this
change shouldn't break anyone's existing workflows for Collabora Online

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Idf4a8b0e94d8720f24c3c4d042a3421f67d22dba