Fix "following" disappearing on user list updates
Previously, following was done by adding a "selected user" class in the
list. This is great, except we are now recreating users at several
points, which loses any classes that we don't know about. This commit
adds a new "following" property to user and when users are created uses
it to inform us if the class should be added

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ia8e03ad75376a39e36911d3a24cb293d2cc15aa8