Convert Control.UserList.js to typescript
I'm working on a set of patches to improve the userlist/following UI in
Collabora Online. Using typescript will allow me to be more certain
about the code I'm writing and make my code easier for other people to
understand.

Additionally, as I'm about to make significant changes to this file,
updating it to typescript now won't even damage the ability to blame too
much!

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Iccf64a01a465d895cbc1c24ad107f85665b5fbf3