Improve styling for the new userlist and following
Previously the userlist and following worked, but were out-of-place
with the rest of the application. This was particularly noticable on
hover where I used a changing border color and text decoration rather
than the usual background change. Additionally, as I was using border
colors to indicate hover rather than a background change this caused
contrast issues on the light theme.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I79373227d6efd397691038ab290a85b2913e0fad