Add chip for "follow the editor"
Previously the improvements to following were limited to only following
different users. We have another way to follow people, though, the
button labelled "always follow the editor".

We can repurpose the existing following chip to support this option by
changing the top text to say you are following the editor rather than
naming a specific user.

Additionally, to avoid inconsistencies with the "following" label on
users it's better to check which user we are following when rendering
the user list rather than keeping track via a "following" boolean on
each user. This boolean was getting out of sync when using the following
the editor feature.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I2b76c1b1e4724cc2aaddbf8fd10b6c4e9155a169