Change convoluted url and name in windowed presentation
Open an empty popup with a blank url, then set the
content of the popup to be our slideURL wrapped inside an
iframe with a user friendly title.
Focus the iframe on load so that arrowkey navigation works
and forward the eventlisterner of the iframe to the parent
window to watch for escape key calls

Adapt patch to LibreOffice Online.

Signed-off-by: NickWingate <nick.wingate@collabora.com>
Change-Id: Ie1063095db51119fd45453b8b4b180fcacb6d7c1